Separate files for both policy types
