+++ lib/apiv2/policy.go
@@ -15,53 +15,8 @@
 package apiv2
 
 import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
+	"github.com/projectcalico/libcalico-go/lib/numorstring"
 )
-
-const (
-	KindNetworkPolicy           = "NetworkPolicy"
-	KindNetworkPolicyList       = "NetworkPolicyList"
-	KindGlobalNetworkPolicy     = "GlobalNetworkPolicy"
-	KindGlobalNetworkPolicyList = "GlobalNetworkPolicyList"
-)
-
-// GlobalNetworkPolicy contains information about a security Policy resource.  This contains a set of
-// security rules to apply.  Security policies allow a selector-based security model which can override
-// the security profiles directly referenced by an endpoint.
-//
-// Each policy must do one of the following:
-//
-//  	- Match the packet and apply an “allow” action; this immediately accepts the packet, skipping
-//        all further policies and profiles. This is not recommended in general, because it prevents
-//        further policy from being executed.
-// 	- Match the packet and apply a “deny” action; this drops the packet immediately, skipping all
-//        further policy and profiles.
-// 	- Fail to match the packet; in which case the packet proceeds to the next policy. If there
-// 	  are no more policies then the packet is dropped.
-//
-// Calico implements the security policy for each endpoint individually and only the policies that
-// have matching selectors are implemented. This ensures that the number of rules that actually need
-// to be inserted into the kernel is proportional to the number of local endpoints rather than the
-// total amount of policy.
-//
-// GlobalNetworkPolicy is globally-scoped (i.e. not Namespaced).
-type GlobalNetworkPolicy struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard object's metadata.
-	Metadata metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of the Policy.
-	Spec PolicySpec `json:"spec,omitempty"`
-}
-
-// NetworkPolicy is the Namespaced-equivalent of the GlobalNetworkPolicy.
-type NetworkPolicy struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard object's metadata.
-	Metadata metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of the Policy.
-	Spec PolicySpec `json:"spec,omitempty"`
-}
 
 // PolicySpec contains the specification for a selector-based security Policy resource.
 type PolicySpec struct {
@@ -135,100 +90,106 @@ const (
 	PolicyTypeEgress  PolicyType = "Egress"
 )
 
-// GlobalNetworkPolicyList contains a list of GlobalNetworkPolicy resources.
-type GlobalNetworkPolicyList struct {
-	metav1.TypeMeta `json:",inline"`
-	Metadata        metav1.ListMeta       `json:"metadata"`
-	Items           []GlobalNetworkPolicy `json:"items"`
+// A Rule encapsulates a set of match criteria and an action.  Both selector-based security Policy
+// and security Profiles reference rules - separated out as a list of rules for both
+// ingress and egress packet matching.
+//
+// Each positive match criteria has a negated version, prefixed with ”Not”. All the match
+// criteria within a rule must be satisfied for a packet to match. A single rule can contain
+// the positive and negative version of a match and both must be satisfied for the rule to match.
+type Rule struct {
+	Action Action `json:"action" validate:"action"`
+	// IPVersion is an optional field that restricts the rule to only match a specific IP
+	// version.
+	IPVersion *int `json:"ipVersion,omitempty" validate:"omitempty,ipversion"`
+	// Protocol is an optional field that restricts the rule to only apply to traffic of
+	// a specific IP protocol. Required if any of the EntityRules contain Ports
+	// (because ports only apply to certain protocols).
+	//
+	// Must be one of these string values: "tcp", "udp", "icmp", "icmpv6", "sctp", "udplite"
+	// or an integer in the range 1-255.
+	Protocol *numorstring.Protocol `json:"protocol,omitempty" validate:"omitempty"`
+	// ICMP is an optional field that restricts the rule to apply to a specific type and
+	// code of ICMP traffic.  This should only be specified if the Protocol field is set to
+	// "icmp" or "icmpv6".
+	ICMP *ICMPFields `json:"icmp,omitempty" validate:"omitempty"`
+	// NotProtocol is the negated version of the Protocol field.
+	NotProtocol *numorstring.Protocol `json:"notProtocol,omitempty" validate:"omitempty"`
+	// NotICMP is the negated version of the ICMP field.
+	NotICMP *ICMPFields `json:"notICMP,omitempty" validate:"omitempty"`
+	// Source contains the match criteria that apply to source entity.
+	Source EntityRule `json:"source,omitempty" validate:"omitempty"`
+	// Destination contains the match criteria that apply to destination entity.
+	Destination EntityRule `json:"destination,omitempty" validate:"omitempty"`
 }
 
-// NewGlobalNetworkPolicy creates a new (zeroed) GlobalNetworkPolicy struct with the TypeMetadata initialised to the current
-// version.
-func NewGlobalNetworkPolicy() *GlobalNetworkPolicy {
-	return &GlobalNetworkPolicy{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       KindGlobalNetworkPolicy,
-			APIVersion: GroupVersionCurrent,
-		},
-	}
+// ICMPFields defines structure for ICMP and NotICMP sub-struct for ICMP code and type
+type ICMPFields struct {
+	// Match on a specific ICMP type.  For example a value of 8 refers to ICMP Echo Request
+	// (i.e. pings).
+	Type *int `json:"type,omitempty" validate:"omitempty,gte=0,lte=254"`
+	// Match on a specific ICMP code.  If specified, the Type value must also be specified.
+	// This is a technical limitation imposed by the kernel’s iptables firewall, which
+	// Calico uses to enforce the rule.
+	Code *int `json:"code,omitempty" validate:"omitempty,gte=0,lte=255"`
 }
 
-// NewGlobalNetworkPolicyList creates a new (zeroed) GlobalNetworkPolicyList struct with the TypeMetadata initialised to the current
-// version.
-func NewGlobalNetworkPolicyList() *GlobalNetworkPolicyList {
-	return &GlobalNetworkPolicyList{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       KindGlobalNetworkPolicyList,
-			APIVersion: GroupVersionCurrent,
-		},
-	}
+// An EntityRule is a sub-component of a Rule comprising the match criteria specific
+// to a particular entity (that is either the source or destination).
+//
+// A source EntityRule matches the source endpoint and originating traffic.
+// A destination EntityRule matches the destination endpoint and terminating traffic.
+type EntityRule struct {
+	// Tag is an optional field that restricts the rule to only apply to traffic that
+	// originates from (or terminates at) endpoints that have profiles with the given tag
+	// in them.
+	Tag string `json:"tag,omitempty" validate:"omitempty,tag"`
+	// Nets is an optional field that restricts the rule to only apply to traffic that
+	// originates from (or terminates at) IP addresses in any of the given subnets.
+	Nets []string `json:"nets,omitempty" validate:"omitempty,dive,cidr"`
+	// Selector is an optional field that contains a selector expression (see Policy for
+	// sample syntax).  Only traffic that originates from (terminates at) endpoints matching
+	// the selector will be matched.
+	//
+	// Note that: in addition to the negated version of the Selector (see NotSelector below), the
+	// selector expression syntax itself supports negation.  The two types of negation are subtly
+	// different. One negates the set of matched endpoints, the other negates the whole match:
+	//
+	//	Selector = "!has(my_label)" matches packets that are from other Calico-controlled
+	// 	endpoints that do not have the label “my_label”.
+	//
+	// 	NotSelector = "has(my_label)" matches packets that are not from Calico-controlled
+	// 	endpoints that do have the label “my_label”.
+	//
+	// The effect is that the latter will accept packets from non-Calico sources whereas the
+	// former is limited to packets from Calico-controlled endpoints.
+	Selector string `json:"selector,omitempty" validate:"omitempty,selector"`
+	// Ports is an optional field that restricts the rule to only apply to traffic that has a
+	// source (destination) port that matches one of these ranges/values. This value is a
+	// list of integers or strings that represent ranges of ports.
+	//
+	// Since only some protocols have ports, if any ports are specified it requires the
+	// Protocol match in the Rule to be set to "tcp" or "udp".
+	Ports []numorstring.Port `json:"ports,omitempty" validate:"omitempty,dive"`
+	// NotTag is the negated version of the Tag field.
+	NotTag string `json:"notTag,omitempty" validate:"omitempty,tag"`
+	// NotNets is an optional field that restricts the rule to only apply to traffic that
+	// does not originate from (or terminate at) an IP address in any of the given subnets.
+	NotNets []string `json:"notNets,omitempty" validate:"omitempty,dive,cidr"`
+	// NotSelector is the negated version of the Selector field.  See Selector field for
+	// subtleties with negated selectors.
+	NotSelector string `json:"notSelector,omitempty" validate:"omitempty,selector"`
+	// NotPorts is the negated version of the Ports field.
+	// Since only some protocols have ports, if any ports are specified it requires the
+	// Protocol match in the Rule to be set to "tcp" or "udp".
+	NotPorts []numorstring.Port `json:"notPorts,omitempty" validate:"omitempty,dive"`
 }
 
-// NetworkPolicyList contains a list of NetworkPolicy resources.
-type NetworkPolicyList struct {
-	metav1.TypeMeta `json:",inline"`
-	Metadata        metav1.ListMeta `json:"metadata"`
-	Items           []NetworkPolicy `json:"items"`
-}
+type Action string
 
-// NewNetworkPolicy creates a new (zeroed) NetworkPolicy struct with the TypeMetadata initialised to the current
-// version.
-func NewNetworkPolicy() *NetworkPolicy {
-	return &NetworkPolicy{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       KindNetworkPolicy,
-			APIVersion: GroupVersionCurrent,
-		},
-	}
-}
-
-// NewNetworkPolicyList creates a new (zeroed) NetworkPolicyList struct with the TypeMetadata initialised to the current
-// version.
-func NewNetworkPolicyList() *NetworkPolicyList {
-	return &NetworkPolicyList{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       KindNetworkPolicyList,
-			APIVersion: GroupVersionCurrent,
-		},
-	}
-}
-
-// GetObjectKind returns the kind of this object.  Required to satisfy Object interface
-func (e *GlobalNetworkPolicy) GetObjectKind() schema.ObjectKind {
-	return &e.TypeMeta
-}
-
-// GetObjectMeta returns the object metadata of this object. Required to satisfy ObjectMetaAccessor interface
-func (e *GlobalNetworkPolicy) GetObjectMeta() metav1.Object {
-	return &e.Metadata
-}
-
-// GetObjectKind returns the kind of this object. Required to satisfy Object interface
-func (el *GlobalNetworkPolicyList) GetObjectKind() schema.ObjectKind {
-	return &el.TypeMeta
-}
-
-// GetListMeta returns the list metadata of this object. Required to satisfy ListMetaAccessor interface
-func (el *GlobalNetworkPolicyList) GetListMeta() metav1.List {
-	return &el.Metadata
-}
-
-// GetObjectKind returns the kind of this object.  Required to satisfy Object interface
-func (e *NetworkPolicy) GetObjectKind() schema.ObjectKind {
-	return &e.TypeMeta
-}
-
-// GetObjectMeta returns the object metadata of this object. Required to satisfy ObjectMetaAccessor interface
-func (e *NetworkPolicy) GetObjectMeta() metav1.Object {
-	return &e.Metadata
-}
-
-// GetObjectKind returns the kind of this object. Required to satisfy Object interface
-func (el *NetworkPolicyList) GetObjectKind() schema.ObjectKind {
-	return &el.TypeMeta
-}
-
-// GetListMeta returns the list metadata of this object. Required to satisfy ListMetaAccessor interface
-func (el *NetworkPolicyList) GetListMeta() metav1.List {
-	return &el.Metadata
-}
+const (
+	Allow Mode = "Allow"
+	Deny       = "Deny"
+	Log        = "Log"
+	Pass       = "Pass"
+)
